% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Data preprocessing for scDecouple analysis.}
\usage{
data_preprocessing(Y, Z)
}
\arguments{
\item{Y}{The perturbation group cell-by-gene matrix.}

\item{Z}{The control group cell-by-gene matrix.}
}
\value{
A list containing the following preprocessed data:
  \itemize{
    \item Z.norm.log.sub: Normalized and log-transformed data matrix for the control group with variable genes.
    \item Y.norm.log.sub: Normalized and log-transformed data matrix for the perturbation group with variable genes.
    \item genes.variable: List of variable genes selected for analysis.
    \item Z.pca.results: PCA results for the control group data.
  }
}
\description{
This function performs data preprocessing for the scDecouple analysis on two cell-by-gene matrices, Y (perturbation group) and Z (control group).
}
\examples{
# Example usage:
preprocessed_data <- data_preprocessing(Y, Z)

}
