% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{downstream_analysis}
\alias{downstream_analysis}
\title{Perform downstream analysis to assess the cellular response of each gene and perform GO enrichment.}
\usage{
downstream_analysis(
  Z.pca.rotation,
  beta.PC.scDecouple,
  Z.norm.log.sub,
  Y.norm.log.sub,
  genes.variable,
  degenes.num = 1500
)
}
\arguments{
\item{Z.pca.rotation}{The matrix of variable loadings containing eigenvectors, gene-by-PC, e.g., "Z.pca.results$rotation" from "data_preprocessing" function.}

\item{beta.PC.scDecouple}{The cellular response of each PC, e.g., "beta.PC.scDecouple" from "scDecouple" function.}

\item{Z.norm.log.sub}{The log-normalized control group gene expression matrix, cell-by-gene, e.g., "Z.norm.log.sub" from "data_preprocessing" function.}

\item{Y.norm.log.sub}{The log-normalized perturbation group gene expression matrix, cell-by-gene, e.g., "Y.norm.log.sub" from "data_preprocessing" function.}

\item{genes.variable}{A list of variable genes selected for analysis.}

\item{degenes.num}{Optional. The number of genes to be included in GO enrichment analysis. Default is 1500.}
}
\value{
A list containing the following results:
  \itemize{
    \item gene.rank: A vector of genes sorted by their response to perturbation.
    \item GO.enrich: GO enrichment results using the enrichr function in the enrichR library.
  }
}
\description{
This function assesses the cellular response of each gene and performs Gene Ontology (GO) enrichment analysis.
}
\examples{
# Example usage:
results <- downstream_analysis(Z.pca.rotation, beta.PC.scDecouple, Z.norm.log.sub, Y.norm.log.sub, genes.variable, degenes.num)

}
