% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta_EM_sep_method.R
\name{get_beta_EM_sep_method}
\alias{get_beta_EM_sep_method}
\title{Perform EM-based deconvolution using the scDecouple method.}
\usage{
get_beta_EM_sep_method(
  y,
  number.component,
  em,
  es,
  el,
  beta_init,
  max.step = 100,
  fix_es = FALSE
)
}
\arguments{
\item{y}{A matrix of observed data, cell-by-PC, from the perturbation group.}

\item{number.component}{The number of components (clusters) to estimate.}

\item{em}{A list of component means for each cluster.}

\item{es}{A list of component covariance matrices for each cluster.}

\item{el}{A vector of component proportions for each cluster.}

\item{beta_init}{The initial estimate of the cellular response.}

\item{max.step}{The maximum number of EM iterations. Default is 100.}

\item{fix_es}{Logical, indicating whether to fix the component covariance matrices. Default is FALSE.}
}
\value{
A list containing the following results:
  \itemize{
    \item beta: The estimated cellular response.
    \item es: A list of updated component covariance matrices.
    \item el: A vector of updated component proportions.
    \item em: A list of updated component means.
    \item Q_after: The log-likelihood value after the EM iterations.
    \item ey: A matrix of component assignment probabilities for each observation.
  }
}
\description{
This function performs EM-based deconvolution to estimate the cellular response using the scDecouple method.
}
\examples{
# Example usage:
result <- get_beta_EM_sep_method(y, number.component, em, es, el, beta_init, max.step, fix_es)

}
