\name{scDecouple}
\alias{scDecouple}
\title{Decoupling cellular response from infected proportion bias in scCRISPR-seq}
\usage{
run_scDecouple(Y,Z)
}
\description{
scDecouple offers a comprehensive suite of functionalities for streamlined scCRISPR-seq data analysis focused on decoupling cellular response from infected proportion bias in scCRISPR-seq. The package has an integrated one-click data pre-processing pipeline, infected proportion bias quantification, cellular response estimation, and downstream analytical tools.
}
\examples{
run_scDecouple(Y,Z)
}
