% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_selection.R
\name{pc_selection}
\alias{pc_selection}
\title{Perform principal component (PC) selection for scDecouple analysis.}
\usage{
pc_selection(
  Z.pca.mat,
  Z.pca.sdev = NULL,
  using_rank = TRUE,
  top.sdev = 20,
  top.multimodal = 3,
  th.sdev = 1,
  th.multimodal = 0.0125
)
}
\arguments{
\item{Z.pca.mat}{The PC matrix of the control group, cells-by-PCs, e.g., result from "data_preprocessing" function.}

\item{Z.pca.sdev}{Optional. The standard deviation of PCs. If not provided, it will be calculated from Z.pca.mat.}

\item{using_rank}{Logical, indicating whether to use rank-based PC selection or not. Default is TRUE.}

\item{top.sdev}{Numeric. When using_rank is TRUE, it sets the threshold of standard deviations of PCs. Default is 20.}

\item{top.multimodal}{Numeric. When using_rank is TRUE, it sets the threshold of dip statistics. Default is 3.}

\item{th.sdev}{Numeric. When using_rank is FALSE, it sets the threshold of standard deviations of PCs. Default is 1.}

\item{th.multimodal}{Numeric. When using_rank is FALSE, it sets the threshold of dip statistics. Default is 0.0125.}
}
\value{
A character vector containing the names of selected PCs.
}
\description{
This function selects principal components (PCs) based on explained variance and multimodality measures.
}
\examples{
# Example usage:
selected_pcs <- pc_selection(Z.pca.mat, Z.pca.sdev, using_rank, top.sdev, top.multimodal, th.sdev, th.multimodal)

}
