% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scDecouple.R
\name{run_scDecouple}
\alias{run_scDecouple}
\title{Run the scDecouple analysis on two cell-by-gene matrices.}
\usage{
run_scDecouple(
  Y,
  Z,
  using_rank = TRUE,
  top.sdev = 30,
  top.multimodal = 3,
  th.sdev = 1,
  th.multimodal = 0.0125,
  degenes.num = 1500,
  seed_ = 0
)
}
\arguments{
\item{Y}{The perturbation group cell-by-gene matrix.}

\item{Z}{The control group cell-by-gene matrix.}

\item{using_rank}{Logical, indicating whether to use rank-based PC selection or not. Default is TRUE.}

\item{top.sdev}{Integer, the threshold of standard deviations of PCs when using_rank is TRUE. Default is 30.}

\item{top.multimodal}{Numeric, the threshold of dip statistics when using_rank is TRUE. Default is 3.}

\item{th.sdev}{Numeric, the threshold of standard deviations of PCs when using_rank is FALSE. Default is 1.}

\item{th.multimodal}{Numeric, the threshold of dip statistics when using_rank is FALSE. Default is 0.0125.}

\item{degenes.num}{Numeric, the number of degenes to calculate GO enrichment. Default is 1500.}

\item{seed_}{Numeric, the random seed for the EM algorithm. Default is 0.}
}
\value{
A list containing the following elements:
  \itemize{
    \item processed_data: Results of data preprocessing.
    \item selected_pcs: Selected principal components.
    \item decouple_results: Results of the scDecouple analysis.
    \item downstream_analysis: Results of downstream analysis.
  }
}
\description{
This function performs the scDecouple analysis on two cell-by-gene matrices, Y (perturbation group) and Z (control group).
}
\examples{
# Example usage:
result <- run_scDecouple(Y, Z)

}
