% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDecouple.R
\name{scDecouple}
\alias{scDecouple}
\title{Perform scDecouple analysis to deconvolute cellular response from infection bias.}
\usage{
scDecouple(Z.pca.mat, Z.pca.rotation, Y.norm.log.sub, select.pcs, seed_ = 0)
}
\arguments{
\item{Z.pca.mat}{The PC matrix of the control group, cell-by-PC, e.g., result from "data_preprocessing" function.}

\item{Z.pca.rotation}{The matrix of variable loadings containing eigenvectors, gene-by-PC, e.g., "Z.pca.results$rotation" from "data_preprocessing" function.}

\item{Y.norm.log.sub}{The PC matrix of the perturbation group, cell-by-PC.}

\item{select.pcs}{A character vector containing the names of selected PCs.}

\item{seed_}{Optional. The random seed for the EM algorithm. Default is 0.}
}
\value{
A list containing the following results:
  \itemize{
    \item ratio.changes: The ratio changes of the perturbation group compared to the control group for each cluster.
    \item beta.PC.scDecouple: The cellular response of each PC to the perturbation.
  }
}
\description{
This function uses scDecouple to deconvolute the cellular response from infection bias in two cell-by-gene matrices.
}
\examples{
# Example usage:
result <- scDecouple(Z.pca.mat, Z.pca.rotation, Y.norm.log.sub, select.pcs, seed_)

}
